advanced_topics/flask_examples/oauth2/consumer_oauth2/requirements.txt
flask==2.3.3
Werkzeug==2.3.7
PyJwt==2.8.0
requests==2.31.0